% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_main.R
\name{confint.fgwqsr}
\alias{confint.fgwqsr}
\title{Generate confidence intervals for group and single pollutant effects through nonparametric bootstrapping.}
\usage{
\method{confint}{fgwqsr}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted object from a fgwqsr() call}

\item{parm}{should be left unspecified}

\item{level}{level of confidence interval to produce -- default is 0.95 for 95\% bootstrap confidence intervals}

\item{...}{optional options -- see details}
}
\description{
Provides confidence intervals for group index, single chemical, and adjusting covariate effects through
 nonparameteric bootstrapping.  Bootstrap samples are stratified by cases and controls, with each bootstrap sample keeping
 same proportions of cases and controls as in the original data.  Each bootstrap sample has the same number of observations
 as the original dataset.  NOTE: Bootstrapped samples are sampled from the original data with replacement!
}
\details{
One can provide the following two arguments:
`boot_reps` - number of bootstrapped replicates
`verbose` - allows messages and progress bars to track progress of bootstrapping.
}
