% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr-package.R
\docType{package}
\name{fgwqsr-package}
\alias{fgwqsr}
\alias{fgwqsr-package}
\title{fgwqsr: Frequentist Grouped Weighted Quantile Sum Regression (FGWQSR)}
\description{
In this package, we have implemented a fully frequentist model fitting to the weighted quantile sum regression problem, which amounts to performing group sign constrained regression. The novelty in our approach is that we are able to perform this model fitting without needing a training and testuing split, while providing both group index and single chemical inferential estimates. The main function in this package is the fgwqsr() function.
}
\keyword{internal}
