% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_updated_source.R
\name{fgwqsr}
\alias{fgwqsr}
\title{Fit a FGWQSR Model}
\usage{
fgwqsr(
  formula,
  data,
  quantiles = 5,
  n_mvn_sims = 10000,
  zero_threshold_cutoff = 0.5,
  verbose = T,
  cores = future::availableCores(),
  optim_control_list = list(maxit = 1000, factr = 0.000000000001, fnscale = 1)
)
}
\arguments{
\item{formula}{A formula for model fitting of FGWQSR.  Please see description for formula construction}

\item{data}{dataframe that contains all covariates and the outcome data.  Column names of dataframe should match those referenced int he model formula.}

\item{quantiles}{number of quantiles to quantize the exposure variables in the mixture portion of the model.  Default value is 5.}

\item{n_mvn_sims}{defines resolution for simulated null distribution for group index and single chemical LRTs.  Default is 10,000.}

\item{zero_threshold_cutoff}{defines a .}

\item{verbose}{Displays messages and progress bar while fitting FGWQSR model.  Default is TRUE.}

\item{cores}{number of cores to parallelize on for fitting nested models and simulated null LRT distributions.  Default is number of available cores on user device.}

\item{optim_control_list}{option to supply control options to optim.}
}
\value{
list with attributes from fgwqsr model fitting.
}
\description{
fgwqsr fits a group signed constrained logistic regression with inference for both group indices and single chemical effects.
The model formula that we pass to fgwqsr() is different than traditional formulas in lm and glm, as we will need to denote our mixture groups.  Three special characters are used in fgwqsr formulas:
 * $|$ - denotes the boundary of a mixture group, used to separate chemicals within a mixture group
 * $/$ - denotes the end of the mixture group specification, adjusting covariates can be added to the formula after this character.  If no adjusting covariates, do not need to specify
 * i. - precedes categorical variables to denote a categorical variable.  For example, if we have a categorical variable cat_var, we would denote this in the model formula by i.cat_var.  This is similar to the stata syntax to declare categorical variables.
}
