% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_updated_source.R
\name{fgwqsr_summary}
\alias{fgwqsr_summary}
\title{Summarize a fgwqsr model fit}
\usage{
fgwqsr_summary(fgwqsr_sol, digits = 6)
}
\arguments{
\item{fgwqsr_sol}{a fitted object from a fgwqsr() call}

\item{digits}{the number of rounding digits to display in summary tables.}
}
\description{
Summarize a fgwqsr model fit
}
