% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_main.R
\name{plot.fgwqsr}
\alias{plot.fgwqsr}
\title{Plot a fgwqsr model fit}
\usage{
\method{plot}{fgwqsr}(object, ...)
}
\arguments{
\item{object}{a fitted object from a fgwqsr() call}
}
\description{
Plots a fgwqsr object.
}
