% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgwqsr.R
\name{plot_betas}
\alias{plot_betas}
\title{Plot Results of Betas}
\usage{
plot_betas(model, filename = NULL, beta_axis_pos = "left")
}
\arguments{
\item{model}{a BGWQSR model object fit from bgwqsr_parallel}

\item{filename}{optional file path to save results of plot in a pdf format}

\item{beta_axis_pos}{choice to put legend for group index plots on the left or right side of plot. Options are "left" or "right".  Default is "left"}
}
\value{
void, prints plots, optionally saves plot in pdf format to filename
}
\description{
#' Plots the 95% credible intervals for each group index effect, along with posterior mean group index estimates, from the BGWQSR posterior distributions.
}
