% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgwqsr_source.R
\name{plotResults}
\alias{plotResults}
\title{Plot Results of Beta and Weight Estimates}
\usage{
plotResults(
  model,
  filename = NULL,
  weight_axis_pos = "left",
  beta_axis_pos = "left"
)
}
\value{
void, prints plots
}
\description{
Plot Results of Beta and Weight Estimates
}
