% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgwqsr_source.R
\name{bgwqsr}
\alias{bgwqsr}
\title{Fit a Bayesian Grouped Weighted Quantile Sum Regression Model}
\usage{
bgwqsr(
  formula,
  data,
  quantiles = 5,
  n.iter = 1000,
  n.burnin = 5000,
  n.thin = 1,
  n.chains = 3,
  n.adapt = 1000,
  inits = NA,
  method = "parallel"
)
}
\arguments{
\item{formula}{A formula for model fitting of BGWQSR.  Please see description for formula construction}

\item{data}{dataframe that contains all covariates and the outcome data.  Column names of dataframe should match those referenced int he model formula.}

\item{quantiles}{number of quantiles to quantize the exposure variables in the mixture portion of the model.  Default value is 5.}

\item{n.iter}{number of mcmc iterations after burnin and adapt iterations PER mcmc chain.}

\item{n.burnin}{number of mcmc burnin samples}

\item{n.thin}{thinning interval for mcmc samples}

\item{n.chains}{number of separate independent mcmc chains to use.}

\item{n.adapt}{number of mcmc samples to perform mcmc adaption.}

\item{inits}{initial values to provide for prior distributions.}

\item{method}{method for which}
}
\value{
list with attributes from fgwqsr model fitting.n
}
\description{
Fits a BGWQSR using the runjags package in parallel over multiple cores
}
