% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgwqsr.R
\name{bgwqsr}
\alias{bgwqsr}
\title{Fit a Bayesian Grouped Weighted Quantile Sum Regression Model for binary outcomes}
\usage{
bgwqsr(
  formula,
  data,
  quantiles = 5,
  n.iter = 10000/n.chains,
  n.burnin = 5000,
  n.thin = 1,
  n.chains = 3,
  n.adapt = 1000,
  inits = NA,
  method = "parallel"
)
}
\arguments{
\item{formula}{A formula for model fitting of BGWQSR.  Please see description for
formula construction}

\item{data}{dataframe that contains all covariates and the outcome data.
Column names of dataframe should match those referenced int he model formula.}

\item{quantiles}{number of quantiles to quantize the exposure variables in the
mixture portion of the model.  Default value is 5.}

\item{n.iter}{number of mcmc iterations after burnin and adapt iterations
PER mcmc chain.}

\item{n.burnin}{number of mcmc burnin samples PER mcmc chain}

\item{n.thin}{thinning interval for mcmc samples PER mcmc chain}

\item{n.chains}{number of separate independent mcmc chains to use.}

\item{n.adapt}{number of mcmc samples to perform mcmc adaption PER mcmc chain.}

\item{inits}{initial values to provide for prior distributions.}

\item{method}{method for mcmc fitting, a passed argument to run.jags function.
Can be one of: rjags, simple, interruptible,
parallel, rjparallel, background, bgparallel, or snow.}
}
\value{
list with attributes from bgwqsr model fitting.
Object from \code{run.jags} function call
}
\description{
Fits a BGWQSR using the runjags package in parallel over multiple cores for binary outcomes.
}
\details{
For additional instructions in using the \code{bgwqsr()} function,
please visit \url{https://github.com/Daniel-Rud/fgwqsr/blob/main/README.md}
}
