% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_main.R
\name{summary.fgwqsr}
\alias{summary.fgwqsr}
\title{Summarize a fgwqsr model fit}
\usage{
\method{summary}{fgwqsr}(object, digits = 6)
}
\arguments{
\item{object}{a fitted object from a fgwqsr() call}

\item{digits}{the number of rounding digits to display in summary tables.}
}
\description{
Summarize a fgwqsr model fit
}
