% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwqsr_main.R
\name{summary.fgwqsr}
\alias{summary.fgwqsr}
\title{Summarize a fgwqsr model fit}
\usage{
\method{summary}{fgwqsr}(object, ...)
}
\arguments{
\item{object}{a fitted object from a fgwqsr() call}

\item{...}{optional options -- see details}
}
\description{
Summarizes a fgwqsr object.
}
\details{
Optional option `digits` to set the number of rounding digits to display in summary tables.
}
